// src/__generated__/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /** A date-time string at UTC, such as 2019-12-03T09:54:33Z, compliant with the date-time format. */
  DateTime: { input: any; output: any; }
  /** The `Upload` scalar type represents a file upload. */
  Upload: { input: any; output: any; }
};
// ...
export type CustomField = {
  __typename?: 'CustomField';
  key: Scalars['String']['output'];
  relationEntities?: Maybe<CustomFieldRelation>;
  value?: Maybe<Scalars['String']['output']>;
};

export type CustomFieldInput = {
  key: Scalars['String']['input'];
  value: Scalars['String']['input'];
};

export type CustomFieldPrivacy = {
  __typename?: 'CustomFieldPrivacy';
  allow: Array<CustomFieldPrivacyOptions>;
};

export type CustomFieldPrivacyInput = {
  allow: Array<CustomFieldPrivacyOptions>;
};

export enum CustomFieldPrivacyOptions {
  Admin = 'ADMIN',
  Own = 'OWN'
}

export type CustomFieldRelation = {
  __typename?: 'CustomFieldRelation';
  medias: Array<Media>;
  members: Array<Member>;
  posts: Array<Post>;
  spaces: Array<Space>;
  tags: Array<Tag>;
};
// ...
export type CustomSeoDetail = {
  __typename?: 'CustomSeoDetail';
  canonicalUrl?: Maybe<Scalars['String']['output']>;
  description?: Maybe<Scalars['String']['output']>;
  noIndex?: Maybe<Scalars['Boolean']['output']>;
  thumbnail?: Maybe<Media>;
  thumbnailId?: Maybe<Scalars['ID']['output']>;
  title?: Maybe<Scalars['String']['output']>;
};
// ...
export type Emoji = {
  __typename?: 'Emoji';
  id: Scalars['ID']['output'];
  text: Scalars['String']['output'];
};
// ...
export type File = {
  __typename?: 'File';
  downloadUrl: Scalars['String']['output'];
  extension: Scalars['String']['output'];
  id: Scalars['ID']['output'];
  name?: Maybe<Scalars['String']['output']>;
  size?: Maybe<Scalars['Int']['output']>;
  status?: Maybe<MediaStatus>;
  url: Scalars['String']['output'];
};
// ...
export type Glyph = {
  __typename?: 'Glyph';
  id: Scalars['ID']['output'];
  text: Scalars['String']['output'];
  variant: GlyphMediaVariant;
};
// ...
export type Image = {
  __typename?: 'Image';
  cropHeight?: Maybe<Scalars['Int']['output']>;
  cropWidth?: Maybe<Scalars['Int']['output']>;
  cropX: Scalars['Int']['output'];
  cropY: Scalars['Int']['output'];
  cropZoom: Scalars['Float']['output'];
  dominantColorHex?: Maybe<Scalars['String']['output']>;
  downloadUrl: Scalars['String']['output'];
  dpi?: Maybe<Scalars['Float']['output']>;
  height?: Maybe<Scalars['Float']['output']>;
  id: Scalars['ID']['output'];
  name?: Maybe<Scalars['String']['output']>;
  status?: Maybe<MediaStatus>;
  url: Scalars['String']['output'];
  urls?: Maybe<MediaUrls>;
  width?: Maybe<Scalars['Float']['output']>;
};
// ...
export type Media = Emoji | File | Glyph | Image;
// ...
export type Member = {
  __typename?: 'Member';
  activeSession?: Maybe<MemberSession>;
  /** @deprecated Has been replaced by locale field */
  attributes: MemberAttributes;
  authMemberProps?: Maybe<MemberAuthMemberProps>;
  badges?: Maybe<Array<MemberBadge>>;
  banner?: Maybe<Media>;
  bannerId?: Maybe<Scalars['ID']['output']>;
  blockedMemberIds?: Maybe<Array<Scalars['String']['output']>>;
  blockedMembers?: Maybe<Array<Member>>;
  createdAt: Scalars['DateTime']['output'];
  /** @deprecated Has been replaced with name */
  displayName?: Maybe<Scalars['String']['output']>;
  email?: Maybe<Scalars['String']['output']>;
  emailStatus?: Maybe<MemberEmailStatus>;
  externalId?: Maybe<Scalars['ID']['output']>;
  externalUrl?: Maybe<Scalars['String']['output']>;
  extraProperties?: Maybe<Array<ExtraProperty>>;
  fields?: Maybe<Array<CustomField>>;
  flagged: Scalars['Boolean']['output'];
  id: Scalars['ID']['output'];
  lastSeen?: Maybe<Scalars['DateTime']['output']>;
  lastSeenAt?: Maybe<Scalars['DateTime']['output']>;
  locale: Scalars['String']['output'];
  name?: Maybe<Scalars['String']['output']>;
  network?: Maybe<Network>;
  networkId: Scalars['ID']['output'];
  newEmail?: Maybe<Scalars['String']['output']>;
  /** @deprecated Use the teammate field instead */
  overrideTeammate?: Maybe<Scalars['Boolean']['output']>;
  profilePicture?: Maybe<Media>;
  profilePictureId?: Maybe<Scalars['ID']['output']>;
  relativeUrl?: Maybe<Scalars['String']['output']>;
  role?: Maybe<Role>;
  roleId: Scalars['ID']['output'];
  score?: Maybe<Scalars['Int']['output']>;
  sessions?: Maybe<Array<MemberSession>>;
  settings?: Maybe<MemberSettings>;
  shortcuts?: Maybe<Array<Shortcut>>;
  spaces?: Maybe<PaginatedSpace>;
  staffReasons?: Maybe<Array<StaffReason>>;
  status: MemberStatus;
  subscribersCount?: Maybe<Scalars['Int']['output']>;
  tagline?: Maybe<Scalars['String']['output']>;
  teammate: Scalars['Boolean']['output'];
  timeZone?: Maybe<Scalars['String']['output']>;
  updatedAt: Scalars['DateTime']['output'];
  url?: Maybe<Scalars['String']['output']>;
  username: Scalars['String']['output'];
  verifiedAt?: Maybe<Scalars['DateTime']['output']>;
};
// ...
export type PageInfo = {
  __typename?: 'PageInfo';
  endCursor?: Maybe<Scalars['String']['output']>;
  hasNextPage: Scalars['Boolean']['output'];
};
// ...
export type PaginatedPost = {
  __typename?: 'PaginatedPost';
  edges?: Maybe<Array<PostEdge>>;
  nodes?: Maybe<Array<Post>>;
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars['Int']['output']>;
};
// ...
export enum PinnedInto {
  Member = 'member',
  Network = 'network',
  Post = 'post',
  Space = 'space'
}
// ...
export type Post = {
  __typename?: 'Post';
  /** @deprecated Has been replaced by allowedReactions. */
  allowedEmojis?: Maybe<Array<Scalars['String']['output']>>;
  allowedReactions?: Maybe<Array<Scalars['String']['output']>>;
  attachmentIds: Array<Scalars['ID']['output']>;
  attachments?: Maybe<Array<File>>;
  authMemberProps?: Maybe<PostAuthMemberProps>;
  contentSummary?: Maybe<PostContentSummary>;
  createdAt: Scalars['DateTime']['output'];
  createdBy?: Maybe<SpaceMember>;
  createdById: Scalars['ID']['output'];
  customSeoDetail?: Maybe<CustomSeoDetail>;
  description?: Maybe<Scalars['String']['output']>;
  embedIds: Array<Scalars['String']['output']>;
  embeds?: Maybe<Array<Embed>>;
  externalId?: Maybe<Scalars['ID']['output']>;
  externalUrl?: Maybe<Scalars['String']['output']>;
  fields?: Maybe<Array<CustomField>>;
  /** @deprecated Use "subscribersCount" instead */
  followersCount?: Maybe<Scalars['Int']['output']>;
  /** @deprecated Has been replaced by allowedReactions. */
  forbiddenEmojis?: Maybe<Array<Scalars['String']['output']>>;
  forbiddenReactions?: Maybe<Array<Scalars['String']['output']>>;
  hasMoreContent: Scalars['Boolean']['output'];
  id: Scalars['ID']['output'];
  imageIds: Array<Scalars['ID']['output']>;
  images?: Maybe<Array<Media>>;
  isAnonymous: Scalars['Boolean']['output'];
  isHidden?: Maybe<Scalars['Boolean']['output']>;
  language?: Maybe<Scalars['String']['output']>;
  lastActivityAt?: Maybe<Scalars['DateTime']['output']>;
  lastActivityDetail?: Maybe<PostLastActivityDetail>;
  /** Whether the post is locked */
  locked: Scalars['Boolean']['output'];
  /** @deprecated This field is going to be deprecated in favor of "fields" */
  mappingFields?: Maybe<Array<PostMappingField>>;
  mentionedMembers: Array<Scalars['String']['output']>;
  mentions?: Maybe<Array<Member>>;
  /** @deprecated There are no negative reactions anymore. */
  negativeReactions?: Maybe<Array<Scalars['String']['output']>>;
  /** @deprecated There are no positive reactions anymore. */
  negativeReactionsCount?: Maybe<Scalars['Int']['output']>;
  networkId: Scalars['ID']['output'];
  owner?: Maybe<SpaceMember>;
  ownerId: Scalars['ID']['output'];
  pinnedInto: Array<PinnedInto>;
  pinnedReplies?: Maybe<Array<Post>>;
  /** @deprecated There are no positive reactions anymore. */
  positiveReactions?: Maybe<Array<Scalars['String']['output']>>;
  /** @deprecated There are no positive reactions anymore. */
  positiveReactionsCount?: Maybe<Scalars['Int']['output']>;
  postType?: Maybe<PostType>;
  postTypeId: Scalars['ID']['output'];
  /** Has been replaced by allowedReactions, but some may still use it */
  primaryReactionType?: Maybe<ReactionType>;
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  reactions?: Maybe<Array<PostReactionDetail>>;
  reactionsCount: Scalars['Int']['output'];
  relativeUrl?: Maybe<Scalars['String']['output']>;
  repliedTo?: Maybe<Post>;
  repliedToId?: Maybe<Scalars['ID']['output']>;
  repliedToIds?: Maybe<Array<Scalars['String']['output']>>;
  repliedTos?: Maybe<Array<Post>>;
  replies?: Maybe<PaginatedPost>;
  repliesCount: Scalars['Int']['output'];
  seoDetail: PostSeoDetail;
  shortContent?: Maybe<Scalars['String']['output']>;
  shortcuts?: Maybe<Array<Shortcut>>;
  singleChoiceReactions?: Maybe<Array<Scalars['String']['output']>>;
  slug?: Maybe<Scalars['String']['output']>;
  space?: Maybe<Space>;
  spaceId: Scalars['ID']['output'];
  status: PostStatus;
  subscribersCount?: Maybe<Scalars['Int']['output']>;
  tagIds?: Maybe<Array<Scalars['String']['output']>>;
  tags?: Maybe<Array<Tag>>;
  textContent?: Maybe<Scalars['String']['output']>;
  thumbnail?: Maybe<Media>;
  thumbnailId?: Maybe<Scalars['String']['output']>;
  title?: Maybe<Scalars['String']['output']>;
  topRepliers: Array<PostTopReplier>;
  /** @deprecated topicIds is replaced by tagIds. */
  topicIds?: Maybe<Array<Scalars['String']['output']>>;
  totalRepliesCount: Scalars['Int']['output'];
  updatedAt: Scalars['DateTime']['output'];
  url?: Maybe<Scalars['String']['output']>;
};
// ...
export type PostAuthMemberProps = {
  __typename?: 'PostAuthMemberProps';
  availableReplyTypes?: Maybe<Array<PostType>>;
  canReact?: Maybe<Scalars['Boolean']['output']>;
  context: PermissionContext;
  /** @deprecated Use "subscribed" instead */
  memberPostNotificationSettingsEnabled?: Maybe<Scalars['Boolean']['output']>;
  permissions?: Maybe<Array<ActionPermissions>>;
  scopes?: Maybe<Array<Scalars['String']['output']>>;
  subscribed?: Maybe<Scalars['Boolean']['output']>;
};

export type PostContentSummary = {
  __typename?: 'PostContentSummary';
  hasMoreContent: Scalars['Boolean']['output'];
  summary: Scalars['String']['output'];
};

export type PostEdge = {
  __typename?: 'PostEdge';
  cursor: Scalars['String']['output'];
  node: Post;
};
// ...
export type PostLastActivityDetail = {
  __typename?: 'PostLastActivityDetail';
  entity?: Maybe<PostLastActivityEntity>;
  entityId: Scalars['ID']['output'];
  entityType: PostLastActivityEntityType;
};
// ...
export type PostReactionDetail = {
  __typename?: 'PostReactionDetail';
  count: Scalars['Int']['output'];
  participants?: Maybe<PaginatedPostReactionParticipant>;
  reacted: Scalars['Boolean']['output'];
  reaction: Scalars['String']['output'];
};
// ...
export type Query = {
// ...
  memberPosts: PaginatedPost;
// ...
};
// ...
export type SpaceMember = {
  __typename?: 'SpaceMember';
  member?: Maybe<Member>;
  role?: Maybe<SpaceRole>;
  space?: Maybe<Space>;
};
// ...
export type Tag = {
  __typename?: 'Tag';
  description?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  slug: Scalars['String']['output'];
  /** @deprecated Tags are network-level objects. spaceId is no longer required to identify them. */
  spaceId?: Maybe<Scalars['String']['output']>;
  title: Scalars['String']['output'];
};
// ...
export type GetMemberPostQuery = { __typename?: 'Query', memberPosts: { __typename?: 'PaginatedPost', edges?: Array<{ __typename?: 'PostEdge', cursor: string, node: { __typename?: 'Post', id: string, title?: string | null, description?: string | null, publishedAt?: any | null, reactions?: Array<{ __typename?: 'PostReactionDetail', count: number, reacted: boolean }> | null, owner?: { __typename?: 'SpaceMember', member?: { __typename?: 'Member', name?: string | null, profilePicture?: { __typename?: 'Emoji' } | { __typename?: 'File' } | { __typename?: 'Glyph' } | { __typename?: 'Image', url: string } | null } | null } | null, tags?: Array<{ __typename?: 'Tag', title: string }> | null, fields?: Array<{ __typename?: 'CustomField', key: string, value?: string | null, relationEntities?: { __typename?: 'CustomFieldRelation', medias: Array<{ __typename?: 'Emoji' } | { __typename?: 'File' } | { __typename?: 'Glyph' } | { __typename?: 'Image', url: string }> } | null }> | null } }> | null } };